requirements.txt
PyYAML==6.0.3
rich==14.3.3
